+++ src/server/sv_main.c
@@ -239,8 +239,7 @@ SV_SendServerCommand(client_t *cl, const qchar *fmt, ...)
   //the actual cause of the bug is probably further downstream and should maybe be addressed later, but this certainly fixes the problem for now
   if (len > 1022)
   {
-    SV_WriteAttackLog(va("SV_SendServerCommand(%d, %.20s...) length %ld > 1022 dropping to avoid buffer overflow\n", ARRAY_INDEX(svs.clients, cl), message, strlen(message)));
-    SV_WriteAttackLog(va("full message [%s]\n", message));
+    SV_WriteAttackLog(va("Warning: q3infoboom/q3msgboom exploit attack from client slot %i.\n", ARRAY_INDEX(svs.clients, cl)));
     return;
   }
 
@@ -732,7 +731,7 @@ SVC_BucketForAddress(const netadr_t *address, const unsigned burst, const unsign
   }
 
   //could not allocate a bucket for this address so write to attack log since it is relevant info
-  SV_WriteAttackLogD(va("SVC_BucketForAddress: couldnt allocate bucket for client %s\n", NET_AdrToString(address)));
+  SV_WriteAttackLogD(va("%s: Could not allocate a bucket for client from %s\n", __func__, NET_AdrToString(address)));
   return NULL;
 }
 
@@ -767,7 +766,7 @@ SVC_RateLimit(rateLimit_t *bucket, const unsigned burst, const unsigned period, 
     return qfalse;
   }
 
-  SV_WriteAttackLogD(va("SVC_RateLimit: burst limit exceeded for bucket %i limit %i\n", bucket->burst, burst));
+  SV_WriteAttackLogD(va("%s: burst limit exceeded for bucket: %i limit: %i\n", __func__, bucket->burst, burst));
   return qtrue;
 }
 
@@ -1062,14 +1061,14 @@ SVC_Status(const netadr_t *from)
   //Chey: bugtraq 12534
   if (!SVC_VerifyChallenge(Cmd_Argv(1)))
   {
-    SV_WriteAttackLog(va("SVC_Status: invalid challenge from %s, dropping request\n", NET_AdrToString(from)));
+    SV_WriteAttackLog(va("%s: Invalid challenge from %s, dropping request.\n", __func__, NET_AdrToString(from)));
     return;
   }
 
   //max challenge of 128
   if (strlen(Cmd_Argv(1)) > 128)
   {
-    SV_WriteAttackLog(va("SVC_Status: challenge length exceeded from %s, dropping request\n", NET_AdrToString(from)));
+    SV_WriteAttackLog(va("%s: Challenge length exceeded from %s, dropping request.\n", __func__, NET_AdrToString(from)));
     return;
   }
 
@@ -1128,7 +1127,7 @@ SVC_Info(const netadr_t *from)
   //Chey: bugtraq 12534
   if (!SVC_VerifyChallenge(Cmd_Argv(1)))
   {
-    SV_WriteAttackLog(va("svc info invalid challenge from %s dropping request\n", NET_AdrToString(from)));
+    SV_WriteAttackLog(va("%s: Invalid challenge from %s, dropping request.\n", __func__, NET_AdrToString(from)));
     return;
   }
 
@@ -1140,7 +1139,7 @@ SVC_Info(const netadr_t *from)
   //a maximum challenge length of 128 should be more than plenty
   if (strlen(Cmd_Argv(1)) > 128)
   {
-    SV_WriteAttackLog(va("svc info challenge length exceeded from %s dropping request\n", NET_AdrToString(from)));
+    SV_WriteAttackLog(va("%s: Challenge length exceeded from %s, dropping request.\n", __func__, NET_AdrToString(from)));
     return;
   }
 
@@ -1364,7 +1363,7 @@ SVC_CheckDRDoS(const netadr_t *from)
 
       if (!ban->flood && ((svs.time - ban->time) >= 3000) && ban->count <= 5)
       {
-        SV_WriteAttackLog(va("unban info flood protect for address %s not flooding\n", NET_AdrToString(from)));
+        SV_WriteAttackLog(va("%s: Unban info flood protect for address %s, they're not flooding.\n", __func__, NET_AdrToString(from)));
         Com_Memset(ban, 0, sizeof(floodBan_t));
         oldestBan = i;
         break;
@@ -1372,7 +1371,7 @@ SVC_CheckDRDoS(const netadr_t *from)
 
       if (ban->count >= 180)
       {
-        SV_WriteAttackLog(va("renewing info flood ban for address %s received %i getinfo or getstatus requests in %i milliseconds\n", NET_AdrToString(from), ban->count, svs.time - ban->time));
+        SV_WriteAttackLog(va("%s: Renewing info flood ban for address %s, received %i getinfo/getstatus requests in %i milliseconds.\n", __func__, NET_AdrToString(from), ban->count, svs.time - ban->time));
         ban->time = svs.time;
         ban->count = 0;
         ban->flood = qtrue;
@@ -1421,7 +1420,7 @@ SVC_CheckDRDoS(const netadr_t *from)
 
   if (specificCount >= 3) //sent three to ip in last two seconds
   {
-    SV_WriteAttackLog(va("possible drdos from address %s add to temp getinfo and getstatus ban list\n", NET_AdrToString(from)));
+    SV_WriteAttackLog(va("%s: Possible DRDoS attack to address %s, putting into temporary getinfo/getstatus ban list.\n", __func__, NET_AdrToString(from)));
     ban = &svs.infoFloodBans[oldestBan];
     ban->adr = modifiedFrom;
     ban->time = svs.time;
@@ -1441,7 +1440,7 @@ SVC_CheckDRDoS(const netadr_t *from)
 
     if (lastGlobalLogTime + 1000 <= svs.time) //one log per second
     {
-      SV_WriteAttackLog("detected flood of arbitrary getinfo and getstatus connectionless packets\n");
+      SV_WriteAttackLog("%s: Detected flood of arbitrary getinfo/getstatus connectionless packets.\n");
       lastGlobalLogTime = svs.time;
     }
 
@@ -1511,7 +1510,6 @@ SVC_RemoteCommand(const netadr_t *from, msg_t *msg)
   {
     valid = qtrue;
     Com_Printf("rcon from %s: %s\n", NET_AdrToString(from), Cmd_ArgsFrom(2));
-    SV_WriteAttackLog(va("rcon from %s: %s\n", NET_AdrToString(from), Cmd_Argv(2)));
   }
 
   if (strlen(sv_rconLog->string))
@@ -1562,7 +1560,6 @@ SVC_RemoteCommand(const netadr_t *from, msg_t *msg)
   else if (!valid || !valid2)
   {
     Com_Printf("bad rcon password\n");
-    SV_WriteAttackLog(va("bad rcon password from %s\n", NET_AdrToString(from)));
   }
   else
   {
@@ -1710,14 +1707,14 @@ SVC_ConnectionlessPacket(const netadr_t *from, msg_t *msg)
     //this will print every 5 'period' msecs
     if (dropped[cmd] > 0 && lastMsg[cmd] + 5000 < now)
     {
-      SV_WriteAttackLog(va("%s: \"%s\" rate limit exceeded, dropped %d request%s\n", __func__, commands[cmd], dropped[cmd], dropped[cmd] == 1 ? "":"s"));
+      SV_WriteAttackLog(va("%s: \"%s\" rate limit exceeded, dropped %d request%s.\n", __func__, commands[cmd], dropped[cmd], dropped[cmd] == 1 ? "":"s"));
       dropped[cmd] = 0;
       lastMsg[cmd] = now;
     }
 
     if (droppedAdr > 0 && lastMsgAdr + 5000 < now)
     {
-      SV_WriteAttackLog(va("%s: IP rate limit exceeded, dropped %d request%s\n", __func__, droppedAdr, droppedAdr == 1 ? "":"s"));
+      SV_WriteAttackLog(va("%s: IP rate limit exceeded, dropped %d request%s.\n", __func__, droppedAdr, droppedAdr == 1 ? "":"s"));
       droppedAdr = 0;
       lastMsgAdr = now;
     }
@@ -1801,7 +1798,7 @@ SVC_ConnectionlessPacket(const netadr_t *from, msg_t *msg)
       break;
 
     default:
-      SV_WriteAttackLog(va("bad connectionless packet from %s:\n%s\n", NET_AdrToString(from), s)); //changed from com dprintf to print in attack log and do com printf if attack log not set
+      SV_WriteAttackLog(va("Bad connectionless packet from %s:\n%s\n", NET_AdrToString(from), s)); //changed from com dprintf to print in attack log and do com printf if attack log not set
       break;
   }
 }
@@ -2245,7 +2242,7 @@ SV_UpdateQueue(void)
   {
     if (svs.time - lastQueue[i] > 4000)
     {
-      SV_WriteAttackLog(va("queue position %d for %s timed out - %d ms\n", i, NET_AdrToString(&queue[i]), svs.time - lastQueue[i]));
+      SV_WriteAttackLog(va("Queue position %d for %s has timed out! - %d ms\n", i, NET_AdrToString(&queue[i]), svs.time - lastQueue[i]));
       continue;
     }